Added new mock classes that can be used for testing:
```
tests/dev/mock/classes/ReflectedBaseClass.php
tests/dev/mock/classes/ReflectedClass.php
tests/dev/mock/classes/ReflectedSubParentClass.php

```
